tests/wpt/tests/navigation-api/commit-behavior/after-transition-redirect-explicit-commit.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<script>
promise_test(async t => {
  // Wait for after the load event so that the navigation doesn't get converted
  // into a replace navigation.
  await new Promise(resolve => window.onload = () => t.step_timeout(resolve, 0));

  let start_length = navigation.entries().length;
  let start_hash = location.hash;
  navigation.onnavigate = t.step_func(e => {
    e.intercept({
      handler: t.step_func(() => {
        assert_equals(location.hash, start_hash);
        assert_equals(new URL(e.destination.url).hash, "#push");

        e.redirect("#redirect1");
        assert_equals(location.hash, start_hash);
        assert_equals(new URL(e.destination.url).hash, "#redirect1");

        e.redirect("#redirect2");
        assert_equals(location.hash, start_hash);
        assert_equals(new URL(e.destination.url).hash, "#redirect2");

        e.commit();
        assert_equals(location.hash, "#redirect2");
        assert_equals(new URL(e.destination.url).hash, "#redirect2");
      }),
      commit: "after-transition"
    });
  });
  await navigation.navigate("#push").committed;
  assert_equals(location.hash, "#redirect2");
  assert_equals(navigation.entries().length, start_length + 1);
}, "redirect() then commit()");
</script>
</body>
